+++ packages/docusaurus-theme-common/src/utils/skipToContentUtils.tsx
@@ -14,8 +14,11 @@ import {useLocationChange} from './useLocationChange';
  * The id of the element that should become focused on a page
  * that does not have a <main> html tag.
  * Focusing the Docusaurus Layout children is a reasonable fallback.
+ *
+ * __ prefix allows search crawlers (Algolia/DocSearch) to ignore anchors
+ * https://github.com/facebook/docusaurus/issues/8883#issuecomment-1516328368
  */
-export const SkipToContentFallbackId = 'docusaurus_skipToContent_fallback';
+export const SkipToContentFallbackId = '__docusaurus_skipToContent_fallback';
 
 /**
  * Returns the skip to content element to focus when the link is clicked.